Rename `resolution` module to `query`
